Refactor

+++ prime/prime.h
@@ -4,4 +4,4 @@
 #include <inttypes.h>
 
 int is_prime(const int64_t x);
-int next_prime(int64_t x);
+int64_t next_prime(int64_t x);
